pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-smart-compute-orchestrator"
version = "0.1.0"
description = "Intelligent AI workload routing across CPU, GPU, quantized models, and cloud APIs"
requires-python = ">=3.11"
dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    # Data Validation
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    # Cache / Queue
    "redis>=5.0.0",
    # Task Queue
    "celery>=5.4.0",
    "celery[redis]>=5.4.0",
    # System Monitoring
    "psutil>=6.0.0",
    "pynvml>=11.5.0",
    # AI / ML
    "google-genai>=1.0.0",
    "tiktoken>=0.7.0",
    "scikit-learn>=1.5.0",
    "xgboost>=2.1.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "joblib>=1.4.0",
    # Dashboard
    "streamlit>=1.38.0",
    "plotly>=5.24.0",
    # Logging
    "structlog>=24.4.0",
    # Config
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    # Utils
    "python-multipart>=0.0.12",
    "anyio>=4.6.0",
]

[project.optional-dependencies]
llama = [
    "llama-cpp-python>=0.3.0",
]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "locust>=2.31.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.hatch.build.targets.wheel]
packages = ["api", "core", "workers", "models", "infrastructure", "config", "dashboard"]
